Refator: 라우터구조 리팩토링
