+++ index.html
@@ -5,7 +5,7 @@
   <meta charset="UTF-8" />
   <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>위플플 | 여정 공유 서비스</title>
+  <title>위플플 | 여정 공유 플랫폼</title>
 </head>
 
 <body>
