+++ src/App.tsx
@@ -6,6 +6,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { BrowserRouter } from 'react-router-dom';
 import MainRouter from '@router/mainRouter';
+import AuthRouter from '@router/authRouter';
 
 const queryClient = new QueryClient();
 
@@ -17,6 +18,7 @@ const App = () => {
           <ThemeProvider theme={theme}>
             <GlobalStyle />
             <MainRouter />
+            <AuthRouter />
           </ThemeProvider>
         </BrowserRouter>
       </RecoilRoot>
