+++ src/router/authRouter.tsx
@@ -1,0 +1,25 @@
+import { Route, Routes } from 'react-router-dom';
+import { useLocation } from 'react-router-dom';
+import { useEffect } from 'react';
+import Signup from '@pages/signup/signup.page';
+import Signin from '@pages/signin/signin.page';
+import MainLayout from './routerLayout';
+
+const AuthRouter = () => {
+  const location = useLocation();
+
+  useEffect(() => {
+    window.scrollTo(0, 0);
+  }, [location]);
+
+  return (
+    <Routes>
+      <Route path="/" element={<MainLayout />}>
+        <Route path="/signup" element={<Signup />} />
+        <Route path="/signin" element={<Signin />} />
+      </Route>
+    </Routes>
+  );
+};
+
+export default AuthRouter;
